Update SseItem's ctor eventType argument to be nullable
