+++ src/libraries/System.Net.ServerSentEvents/tests/System.Net.ServerSentEvents.Tests.csproj
@@ -5,11 +5,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Include="SseItemTests.cs" />
     <Compile Include="SseParserTests.cs" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\System.Net.ServerSentEvents.csproj"/>
+    <ProjectReference Include="..\src\System.Net.ServerSentEvents.csproj" />
     <ProjectReference Include="$(LibrariesProjectRoot)System.Text.Json\gen\System.Text.Json.SourceGeneration.Roslyn4.4.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" SetTargetFramework="TargetFramework=netstandard2.0" />
   </ItemGroup>
